Ajuste controller usuário

+++ src/main/java/br/com/campushop/campushop_backend/repository/UsuarioRepository.java
@@ -1,13 +1,15 @@
 package br.com.campushop.campushop_backend.repository;
 
 import br.com.campushop.campushop_backend.model.Usuario;
+import jakarta.persistence.criteria.CriteriaBuilder.In;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
     
     Optional<Usuario> findByEmail(String email);
